% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bea.R
\name{bea_datasets}
\alias{bea_datasets}
\title{Get Available Datasets from BEA API}
\usage{
bea_datasets(api_key = NULL)
}
\arguments{
\item{api_key}{Your API key for the BEA API. If NULL, the function will attempt to retrieve the API key from the environment variable BEA_API_KEY.}
}
\value{
A data frame listing all datasets available from the BEA API.
}
\description{
Retrieves a list of datasets available from the BEA API. It constructs a query with minimal parameters to fetch dataset names.
}
\examples{
datasets <- bea_datasets()
}
