% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bea.R
\name{bea_get_data}
\alias{bea_get_data}
\title{Fetch Data from BEA API}
\usage{
bea_get_data(params, api_key = NULL)
}
\arguments{
\item{params}{List of parameters for the BEA API request.
This should include at least `UserID`, `method`, `datasetname`,
`KeyCode`, `GeoFips`, `Year`, and `ResultFormat`.}
}
\value{
A data frame containing the requested data if the API call is successful.
        If the call fails, it will print an error message and return NULL.
}
\description{
This function retrieves data from the BEA API based on the provided parameters.
It handles the request, checks the response status, and parses the JSON content
into a data frame.
}
\examples{
params <- list(
  datasetname = "GDPbyIndustry",
  Frequency = "A",
  Industry = "31G",
  TableID = "1",
  Year = "ALL"
)
bea_data <- bea_get_data(params)

}
