% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bea.R
\name{bea_parameter_values}
\alias{bea_parameter_values}
\title{Get Parameter values for a Specific BEA Dataset}
\usage{
bea_parameter_values(dataset, parameter, api_key = NULL)
}
\arguments{
\item{dataset}{The name of the dataset for which parameter values are requested.}

\item{parameter}{The name of the parameter for which values are requested.}

\item{api_key}{Your API key for the BEA API. If NULL, the function will attempt to retrieve the API key from  the environment variable BEA_API_KEY.}
}
\value{
A data frame containing the parameters for the specified dataset.
}
\description{
Fetches and lists parameters available for a specific dataset from the BEA API, which are necessary for constructing detailed queries.
}
\examples{
parameter_values <- bea_parameter_values(dataset = "GDPbyIndustry", parameter = "TableID")
}
