% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bea.R
\name{bea_parameters}
\alias{bea_parameters}
\title{Get Parameters for a Specific BEA Dataset}
\usage{
bea_parameters(dataset, api_key = NULL)
}
\arguments{
\item{dataset}{The name of the dataset for which parameters are requested.}

\item{api_key}{Your API key for the BEA API. If NULL, the function will attempt to retrieve the API key from  the environment variable BEA_API_KEY.}
}
\value{
A data frame containing the parameters for the specified dataset.
}
\description{
Fetches and lists parameters available for a specific dataset from the BEA API, which are necessary for constructing detailed queries.
}
\examples{
parameters <- bea_parameters("GDPbyIndustry")
}
