% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dip_plot}
\alias{dip_plot}
\title{Plot diptych}
\usage{
dip_plot(
  synth_obj,
  weight_data,
  nace,
  plot_var,
  base_year,
  high_country,
  high_countries
)
}
\arguments{
\item{synth_obj}{a synth_obj.}

\item{weight_data}{Data with weighted variables.}

\item{nace}{nace classification to use.}

\item{plot_var}{A variable to plot.}

\item{high_country}{A country highlighted by line size.}

\item{high_countries}{Countries highlighted by line colour.}
}
\description{
One large level plot and one difference plots
}
