% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bea.R
\name{get_bea}
\alias{get_bea}
\title{Fetch Data from BEA API}
\usage{
get_bea(params)
}
\arguments{
\item{params}{List of parameters for the BEA API request.
This should include at least `UserID`, `method`, `datasetname`,
`KeyCode`, `GeoFips`, `Year`, and `ResultFormat`.}
}
\value{
A data frame containing the requested data if the API call is successful.
        If the call fails, it will print an error message and return NULL.
}
\description{
This function retrieves data from the BEA API based on the provided parameters.
It handles the request, checks the response status, and parses the JSON content
into a data frame.
}
\examples{
params <- list(
  UserID = "YourAPIKeyHere",
  method = "GETDATA",
  datasetname = "RegionalData",
  KeyCode = "PCPI_CI",
  GeoFips = "STATE",
  Year = "2020",
  ResultFormat = "JSON"
)
data <- get_bea(params)

}
