% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{prod_ind_plot_high}
\alias{prod_ind_plot_high}
\title{Productivity index plot with highlighting}
\usage{
prod_ind_plot_high(data, plot_var, base_year, high_country, high_countries)
}
\arguments{
\item{data}{A data to plot.}

\item{high_country}{A country highlighted by line size.}

\item{high_countries}{Countries highlighted by line colour.}
}
\description{
Productivity index plot with highlighting
}
