% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{rebase}
\alias{rebase}
\title{Rebase (or base) index}
\usage{
rebase(x, time, baseyear)
}
\arguments{
\item{x}{a numeric vector. An index to rebase}

\item{time}{a time variable in a Date format.}

\item{baseyear}{a year or vector of years.}
}
\description{
Rebase (or base) index
}
