% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{rel_one_plot}
\alias{rel_one_plot}
\title{Weighted plot of variables}
\usage{
rel_one_plot(
  .data,
  nace,
  w_plot_var = c(`Työn tuottavuus` = "lp_ind"),
  title = "Suomi suhteessa euroalueeseen ja\\nEurooppalaisiin maihin kauppapainoin"
)
}
\arguments{
\item{.data}{A data to plot.}

\item{nace}{A industry classification to plot}
}
\description{
Weighted plot of variables
}
