% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_figs.R
\name{save_figs}
\alias{save_figs}
\title{Save figures for report}
\usage{
save_figs(
  filename,
  plot = last_plot(),
  width = 13.5,
  height = 13.5,
  en = TRUE,
  fi = TRUE,
  translated = FALSE
)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{width}{Plot size}

\item{height}{Plot size}
}
\description{
Uses ggsave.
}
