% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_figs.R
\name{save_trip_plot_data}
\alias{save_trip_plot_data}
\title{Save data from triplot to csv-files}
\usage{
save_trip_plot_data(filename, plot = last_plot())
}
\description{
Save data from triplot to csv-files
}
