% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_funs.R
\name{sc_mean}
\alias{sc_mean}
\title{Synthetic with mean}
\usage{
sc_mean(.data, value_var, i_time)
}
\description{
Synthetic with mean
}
\examples{
sc_mean(filter(synth_dat10, nace0 == "private"), value_var = "lp_ind", 2008)
}
