% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_board_theme}
\alias{set_board_theme}
\title{Set Competetive Board theme}
\usage{
set_board_theme(base_size = 11)
}
\arguments{
\item{base_size}{A font base size.}
}
\description{
Set Competetive Board theme
}
