% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssca_helper.R
\name{ssca_est_loop}
\alias{ssca_est_loop}
\title{Loop SSCA estimations}
\usage{
ssca_est_loop(.value_var, .data, ...)
}
\arguments{
\item{...}{Further arguments passed to \link{\code[ssca_est]}.}

\item{value_var}{A string vector of names of the variable to estimate over.}

\item{data}{A data.frame}
}
\description{
Loop SSCA estimations
}
