% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssca_plot.R
\name{ssca_plot}
\alias{ssca_plot}
\title{Plot ssca object}
\usage{
ssca_plot(
  title,
  ssca_obj,
  legend_arg = c("Suomi", "Synteettinen kontrolli", "Yksi maa jätetty pois maajoukosta")
)
}
\arguments{
\item{legend_arg}{}
}
\description{
Plot ssca object
}
