% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{synt_plot}
\alias{synt_plot}
\title{Plot syntetic control figures}
\usage{
synt_plot(ssca_obj, nace, plot_var, base_year, high_country, high_countries)
}
\arguments{
\item{ssca_obj}{A ssca_obj.}

\item{nace}{A nace classification to use.}

\item{plot_var}{A variable to plot.}

\item{high_country}{A country highlighted by line size.}

\item{high_countries}{Countries highlighted by line colour.}
}
\description{
two ssca plots
}
