% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_funs.R
\name{synth_est_loop}
\alias{synth_est_loop}
\title{Loop SSCA estimations}
\usage{
synth_est_loop(.value_var, .data, i_time)
}
\arguments{
\item{value_var}{A string vector of names of the variable to estimate over.}

\item{data}{A data.frame}

\item{...}{Further arguments passed to \link{\code[ssca_est]}.}
}
\description{
Loop SSCA estimations
}
