% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_plot.R
\name{synth_plot_data}
\alias{synth_plot_data}
\title{Get plot data from synth_obj}
\usage{
synth_plot_data(synth_obj, plot_var = NULL, nace = NULL)
}
\arguments{
\item{synth_obj}{a synth_obj.}

\item{plot_var}{A name of variable to plot.}

\item{nace}{A name of the nace0 classification to plot.}
}
\description{
Get plot data from synth_obj
}
\examples{
  synth_plot_data(synth_obj = synth10_est_results, plot_var = "lp_ind", nace = "private")
}
