% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{prod_ind_plot_high}
\alias{prod_ind_plot_high}
\title{Productivity index plot with highlighting}
\usage{
prod_ind_plot_high(data)
}
\arguments{
\item{data}{a data to plot}
}
\description{
Productivity index plot with highlighting
}
